partial rollback
