+++ src/main/java/eu/dissco/annotationprocessingservice/database/jooq/DefaultCatalog.java
@@ -6,7 +6,6 @@ package eu.dissco.annotationprocessingservice.database.jooq;
 
 import java.util.Arrays;
 import java.util.List;
-
 import org.jooq.Constants;
 import org.jooq.Schema;
 import org.jooq.impl.CatalogImpl;
