+++ src/main/java/eu/dissco/annotationprocessingservice/database/jooq/Indexes.java
@@ -6,7 +6,6 @@ package eu.dissco.annotationprocessingservice.database.jooq;
 
 import eu.dissco.annotationprocessingservice.database.jooq.tables.Annotation;
 import eu.dissco.annotationprocessingservice.database.jooq.tables.MasJobRecord;
-
 import org.jooq.Index;
 import org.jooq.OrderField;
 import org.jooq.impl.DSL;
